requirements.txt
pymupdf
pillow
pytesseract
deep-translator
python-dotenv

chromadb
sentence-transformers
fastapi
uvicorn
streamlit
requests